Fix /etc/hosts auto-populate task
The way it worked before was fundamentally broken. As soon as an
inventory would contain more than one host, we would bind ip addresses
to the wrong fqdn.

With this commit, we lose the ability to populate our host with our
whole inventory, but at least it's not broken.

fixes #10