Switch from "include:" to "import_tasks:" to resolve warnings
As of Ansible 2.4, "include:" is deprecated in favor of
"include_tasks:" and "import_tasks:". This change removes the
DEPRECATION WARNING that recent Ansible versions output as a result.

As a side-effect, this role no longer works with Ansible versions
older than 2.4, but older versions of this role (such as v1.2) are
still available.

Signed-off-by: toby cabot <toby@caboteria.org>